Split MSSQLSourceTable test cases to own test file and cleaned up CaptureReconciler tests
