+++ modules/module-mssql/test/src/CaptureReconciler.test.ts
@@ -1,15 +1,10 @@
 import { CaptureInstance } from '@module/common/CaptureInstance.js';
-import { LSN } from '@module/common/LSN.js';
-import { MSSQLSourceTable } from '@module/common/MSSQLSourceTable.js';
-import {
-  createCaptureReconciler,
-  MSSQLSourceMetadata,
-  readCaptureMetadata
-} from '@module/replication/CaptureReconciler.js';
+import { createCaptureReconciler, readCaptureMetadata } from '@module/replication/CaptureReconciler.js';
 import type { MSSQLTableReconciliationContext } from '@module/replication/MSSQLTableReconciliationContext.js';
 import { MSSQLTableReconciliationState } from '@module/replication/MSSQLTableReconciliationContext.js';
 import { SourceEntityDescriptor, SourceTable } from '@powersync/service-core';
 import { describe, expect, it } from 'vitest';
+import { createCaptureInstance, createSourceDescriptor, createSourceTableCandidate } from './util.js';
 
 function reconcile(context: MSSQLTableReconciliationContext, candidates: SourceTable[]) {
   return createCaptureReconciler(context)({ source: context.source, candidates });
@@ -140,9 +135,7 @@ describe('createCaptureReconciler', () => {
       reconcile(readyContext([createCaptureInstance(50)], changedSource), [
         createSourceTableCandidate('old', { captureTableObjectId: 40 })
       ])
-    ).toThrow(
-      /Table \[dbo\]\.\[users\] no longer matches the source table binding.*already-replicated data is retained/
-    );
+    ).toThrow(/Table \[dbo]\.\[users] no longer matches the source table binding.*already-replicated data is retained/);
   });
 
   it('drops stale incompatible candidates when a compatible candidate anchors the binding', () => {
@@ -158,81 +151,6 @@ describe('createCaptureReconciler', () => {
     expect(resolution.incompatibleTables.map((table) => table.id)).toEqual(['mismatch']);
   });
 });
-
-describe('MSSQLSourceTable.setCaptureInstance', () => {
-  it('sets the instance matching the persisted capture-table object id', () => {
-    const table = new MSSQLSourceTable(createSourceDescriptor(), [
-      createSourceTableCandidate('a', { captureTableObjectId: 40 })
-    ]);
-    const expected = createCaptureInstance(40);
-
-    table.setCaptureInstance([createCaptureInstance(50), expected]);
-
-    expect(table.captureInstance).toBe(expected);
-  });
-
-  it('sets null when the binding is legacy or the pinned instance is unavailable', () => {
-    const legacy = new MSSQLSourceTable(createSourceDescriptor(), [createSourceTableCandidate('legacy')]);
-    const pinned = new MSSQLSourceTable(createSourceDescriptor(), [
-      createSourceTableCandidate('pinned', { captureTableObjectId: 40 })
-    ]);
-
-    legacy.setCaptureInstance([createCaptureInstance(40)]);
-    pinned.setCaptureInstance([createCaptureInstance(40)]);
-    pinned.setCaptureInstance([createCaptureInstance(50)]);
-
-    expect(legacy.captureInstance).toBeNull();
-    expect(pinned.captureInstance).toBeNull();
-  });
-});
-
-/**
- * Create a capture instance with the given capture-table object id.
- */
-function createCaptureInstance(objectId: number): CaptureInstance {
-  return {
-    name: `dbo_users_${objectId}`,
-    objectId,
-    minLSN: LSN.fromString(LSN.ZERO),
-    createDate: new Date(),
-    pendingSchemaChanges: []
-  };
-}
-
-/**
- * Build a persisted source-table candidate with optional capture metadata.
- */
-function createSourceTableCandidate(
-  id: string,
-  metadata?: MSSQLSourceMetadata,
-  overrides: Partial<ConstructorParameters<typeof SourceTable>[0]> = {}
-): SourceTable {
-  return new SourceTable({
-    id,
-    ref: { connectionTag: 'default', schema: 'dbo', name: 'users' },
-    objectId: 100,
-    replicaIdColumns: [{ name: 'id', type: 'int', typeId: 56 }],
-    snapshotComplete: true,
-    bucketDataSources: [],
-    parameterLookupSources: [],
-    sourceMetadata: metadata,
-    ...overrides
-  });
-}
-
-/**
- * Create a source descriptor with optional identity overrides.
- */
-function createSourceDescriptor(overrides: Partial<SourceEntityDescriptor> = {}): SourceEntityDescriptor {
-  return {
-    connectionTag: 'default',
-    schema: 'dbo',
-    name: 'users',
-    objectId: 100,
-    replicaIdColumns: [{ name: 'id', type: 'int', typeId: 56 }],
-    ...overrides
-  };
-}
 
 function readyContext(
   captureInstances: CaptureInstance[],
